refactor (#19)
