+++ backend/src/common.rs
@@ -1,3 +1,4 @@
 pub mod auth;
 pub mod error;
+pub mod repository;
 pub mod response;
