+++ backend/src/common/response.rs
@@ -11,10 +11,10 @@ pub struct ApiResponse<T> {
     pub data: Option<T>,
 }
 
-#[derive(Debug, Serialize)]
-pub struct MessageResponse {
-    pub message: String,
-}
+// #[derive(Debug, Serialize)]
+// pub struct MessageResponse {
+//     pub message: String,
+// }
 
 pub type ApiResult<T> = Result<(StatusCode, Json<ApiResponse<T>>), AppError>;
 pub type ApiCookieResult<T> = Result<(StatusCode, CookieJar, Json<ApiResponse<T>>), AppError>;
@@ -27,10 +27,10 @@ pub fn ok<T: Serialize>(msg: &str, data: T) -> Json<ApiResponse<T>> {
     })
 }
 
-pub fn message(msg: &str) -> Json<ApiResponse<MessageResponse>> {
-    Json(ApiResponse {
-        success: true,
-        message: msg.to_string(),
-        data: None,
-    })
-}
+// pub fn message(msg: &str) -> Json<ApiResponse<MessageResponse>> {
+//     Json(ApiResponse {
+//         success: true,
+//         message: msg.to_string(),
+//         data: None,
+//     })
+// }
