+++ backend/src/modules.rs
@@ -1,1 +1,2 @@
+pub mod auth;
 pub mod users;
