+++ backend/src/modules/auth.rs
@@ -1,3 +1,4 @@
 pub mod dto;
 pub mod handler;
+pub mod model;
 pub mod service;
