--- backend/src/modules/users/model.rs
@@ -1,17 +1,0 @@
-use serde::{Deserialize, Serialize};
-use uuid::Uuid;
-
-#[derive(Debug)]
-pub struct User {
-    pub id: Uuid,
-    pub email: String,
-    pub password_hash: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Claims {
-    pub sub: String,
-    pub email: String,
-    pub exp: usize,
-    pub iat: usize,
-}
